Fixes `mikepb/react-draggable#2`
The previous `package.json` applied a global Browserify transform
without requiring it as a production dependency. This caused Browserify
bundling to fail unless it was installed elsewhere. The use of `require`
was also broken, because `react` was not a production dependency. This
just fixes these minor configuration issues by removing the transform
and making `react` a productiond dependency.